+++ bw.workflow.registerconvictedpresent/src/main/java/sa/gov/nic/bio/bw/workflow/registerconvictedpresent/RegisterConvictedReportPresentWorkflow.java
@@ -93,7 +93,7 @@ public class RegisterConvictedReportPresentWorkflow extends WizardWorkflowBase
 										"registerConvictedReport.fingerprint.acceptBadQualityFingerprintMinRetries"));
 				
 				setData(SlapFingerprintsCapturingFxController.class, "hidePreviousButton", Boolean.TRUE);
-//				setData(SlapFingerprintsCapturingFxController.class, "allow9MissingWithNoRole", Boolean.TRUE);
+				setData(SlapFingerprintsCapturingFxController.class, "allow9MissingWithNoRole", Boolean.TRUE);
 				setData(SlapFingerprintsCapturingFxController.class, "acceptBadQualityFingerprint",
 				        acceptBadQualityFingerprint);
 				setData(SlapFingerprintsCapturingFxController.class, "acceptBadQualityFingerprintMinRetires",