+++ bw.workflow.registeriris/src/main/java/sa/gov/nic/bio/bw/workflow/registeriris/tasks/SubmitIrisRegistrationWorkflowTask.java
@@ -2,7 +2,6 @@ package sa.gov.nic.bio.bw.workflow.registeriris.tasks;
 
 import sa.gov.nic.bio.bw.core.Context;
 import sa.gov.nic.bio.bw.core.workflow.Input;
-import sa.gov.nic.bio.bw.core.workflow.Output;
 import sa.gov.nic.bio.bw.core.workflow.Signal;
 import sa.gov.nic.bio.bw.core.workflow.WorkflowTask;
 import sa.gov.nic.bio.bw.workflow.registeriris.webservice.IrisRegistrationAPI;
@@ -12,15 +11,15 @@ public class SubmitIrisRegistrationWorkflowTask extends WorkflowTask
 	@Input(alwaysRequired = true) private Long personId;
 	@Input(alwaysRequired = true) private String rightIrisBase64;
 	@Input(alwaysRequired = true) private String leftIrisBase64;
-	@Output private Long tcn;
+//	@Output private Long tcn;
 	
 	@Override
 	public void execute() throws Signal
 	{
 		var api = Context.getWebserviceManager().getApi(IrisRegistrationAPI.class);
-		var apiCall = api.submitIrisRegistration(workflowId, workflowTcn, personId, rightIrisBase64, leftIrisBase64);
+		var apiCall = api.replaceIris(workflowId, workflowTcn, personId, rightIrisBase64, leftIrisBase64, null);
 		var taskResponse = Context.getWebserviceManager().executeApi(apiCall);
 		resetWorkflowStepIfNegativeOrNullTaskResponse(taskResponse);
-		tcn = taskResponse.getResult();
+//		tcn = taskResponse.getResult();
 	}
 }